Finishing

+++ app/src/main/java/com/uberalles/whatsappquickchat/ui/history/HistoryPage.kt
@@ -1,6 +1,7 @@
 package com.uberalles.whatsappquickchat.ui.history
 
 import android.content.ActivityNotFoundException
+import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import android.widget.Toast
@@ -17,13 +18,13 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.CornerSize
@@ -45,13 +46,12 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.Alignment.Companion.CenterHorizontally
 import androidx.compose.ui.Alignment.Companion.CenterVertically
-import androidx.compose.ui.Alignment.Companion.End
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Devices
@@ -59,8 +59,15 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
+import com.airbnb.lottie.compose.LottieAnimation
+import com.airbnb.lottie.compose.LottieCompositionSpec
+import com.airbnb.lottie.compose.LottieConstants
+import com.airbnb.lottie.compose.animateLottieCompositionAsState
+import com.airbnb.lottie.compose.rememberLottieComposition
 import com.uberalles.whatsappquickchat.MainViewModel
 import com.uberalles.whatsappquickchat.R
+import com.uberalles.whatsappquickchat.database.History
+import com.uberalles.whatsappquickchat.ui.theme.RobotoSlab
 
 @Composable
 fun HistoryPage(
@@ -69,33 +76,44 @@ fun HistoryPage(
 ) {
     val context = LocalContext.current
     val history by viewModel.getAll().observeAsState(listOf())
-    Scaffold(topBar = { AppBar(viewModel) }) {
+    val emptyList by rememberLottieComposition(spec = LottieCompositionSpec.RawRes(R.raw.empty_list))
+    val progress by animateLottieCompositionAsState(
+        composition = emptyList,
+        iterations = LottieConstants.IterateForever
+    )
+
+    Scaffold(topBar = { AppBar(viewModel, isEmpty = history.isEmpty()) }) {
         Box(modifier = Modifier.padding(it)) {
-            LazyColumn {
-                items(history) { history ->
-                    ListHistory(
-                        phoneNumber = history.phoneNumber,
-                        message = history.message,
-                        onClickSend = {
-                            val url = "https://wa.me/${history.phoneNumber}?text=${history.message}"
-                            val intent = Intent(Intent.ACTION_VIEW).apply {
-                                data =
-                                    Uri.parse(url)
-                                `package` = "com.whatsapp"
-                            }
-                            try {
-                                context.startActivity(intent)
-                            } catch (e: ActivityNotFoundException) {
-                                Toast.makeText(
-                                    context,
-                                    "WhatsApp not installed",
-                                    Toast.LENGTH_SHORT
-                                ).show()
-                            }
-                        }
+            if (history.isEmpty()) {
+                Column(
+                    modifier = Modifier.fillMaxSize(),
+                    verticalArrangement = Arrangement.SpaceEvenly
+                ) {
+                    Text(
+                        fontSize = 24.sp,
+                        fontFamily = RobotoSlab,
+                        fontWeight = FontWeight.W600,
+                        text = "There is nothing here,\ntry to message some numbers first!",
+                        textAlign = TextAlign.Center
                     )
+                    LottieAnimation(composition = emptyList, progress = { progress })
+
+                }
+            } else {
+                LazyColumn {
+                    items(history) { history ->
+                        ListHistory(
+                            phoneNumber = history.phoneNumber,
+                            message = history.message,
+                            createdAt = history.createdAt,
+                            onClickSend = {
+                                sendAgain(context = context, history = history)
+                            }
+                        )
+                    }
                 }
             }
+
         }
 
     }
@@ -106,7 +124,8 @@ fun HistoryPage(
 fun ListHistory(
     phoneNumber: String,
     message: String,
-    onClickSend : () -> Unit
+    createdAt: String,
+    onClickSend: () -> Unit
 ) {
     var expanded by remember { mutableStateOf(false) }
     val transition = updateTransition(targetState = expanded, label = "transition")
@@ -144,18 +163,15 @@ fun ListHistory(
             modifier = Modifier
                 .fillMaxWidth(),
             verticalArrangement = Arrangement.Center,
-            horizontalAlignment = CenterHorizontally
         ) {
             Row(
                 modifier = Modifier
                     .fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween
             ) {
                 Text(
                     modifier = Modifier
                         .padding(8.dp)
-                        .width(250.dp)
-                    ,
+                        .width(250.dp),
                     text = phoneNumber,
                     fontSize = 24.sp,
                     overflow = TextOverflow.Ellipsis,
@@ -169,7 +185,7 @@ fun ListHistory(
                 ) {
                     Icon(
                         imageVector = Icons.Default.Send,
-                        contentDescription = null,
+                        contentDescription = "Message this number",
                         modifier = Modifier
                             .size(30.dp)
                             .rotate(180f)
@@ -180,7 +196,7 @@ fun ListHistory(
                     )
                     Icon(
                         imageVector = Icons.Default.KeyboardArrowUp,
-                        contentDescription = null,
+                        contentDescription = "Expand the description",
                         modifier = Modifier
                             .size(48.dp)
                             .rotate(iconRotation)
@@ -194,13 +210,21 @@ fun ListHistory(
                 enter = enterTransition,
                 exit = exitTransition
             ) {
-                Text(
+                Column(
                     modifier = Modifier
                         .padding(horizontal = 10.dp, vertical = 5.dp),
-                    text = message,
-                    textAlign = TextAlign.Justify,
-                    fontSize = 18.sp
-                )
+                    horizontalAlignment = Alignment.Start
+                ) {
+                    Text(
+                        text = message,
+                        textAlign = TextAlign.Start,
+                        fontSize = 18.sp
+                    )
+                    Spacer(modifier = Modifier.height(10.dp))
+                    Text(
+                        text = createdAt
+                    )
+                }
             }
         }
     }
@@ -209,27 +233,54 @@ fun ListHistory(
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun AppBar(
-    viewModel: MainViewModel
+    viewModel: MainViewModel,
+    isEmpty: Boolean
 ) {
     TopAppBar(
         modifier = Modifier,
-        title = { Text(text = "Recent History") },
+        title = { },
         actions = {
-            IconButton(
-                onClick = {
-                    viewModel.deleteAll()
-                },
-            ) {
-                Icon(
-                    modifier = Modifier
-                        .height(50.dp)
-                        .width(50.dp),
-                    painter = painterResource(id = R.drawable.ic_delete),
-                    contentDescription = null
-                )
+            if (isEmpty) {
+                
+            } else {
+                IconButton(
+                    onClick = {
+                        viewModel.deleteAll()
+                    },
+                ) {
+                    Icon(
+                        modifier = Modifier
+                            .height(50.dp)
+                            .width(50.dp),
+                        painter = painterResource(id = R.drawable.ic_delete),
+                        contentDescription = null
+                    )
+                }
             }
+
         }
     )
+}
+
+private fun sendAgain(
+    context: Context,
+    history: History
+) {
+    val url = "https://wa.me/${history.phoneNumber}?text=${history.message}"
+    val intent = Intent(Intent.ACTION_VIEW).apply {
+        data =
+            Uri.parse(url)
+        `package` = "com.whatsapp"
+    }
+    try {
+        context.startActivity(intent)
+    } catch (e: ActivityNotFoundException) {
+        Toast.makeText(
+            context,
+            "WhatsApp not installed",
+            Toast.LENGTH_SHORT
+        ).show()
+    }
 }
 
 @Preview(showBackground = true, device = Devices.PIXEL_4)
@@ -238,6 +289,8 @@ fun ListPreview() {
     ListHistory(
         phoneNumber = "12345678910111111231241312",
         onClickSend = { },
-        message = "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
+        createdAt = "12/12/2021",
+        message = "e1s2e1 2c4e31 d1234d "
     )
 }
+
